♻️ refactor: 라디오 그룹 내 헬퍼 함수 분리 #330
